+++ Src/System.Linq.Dynamic.dnx.Tests/project.json
@@ -16,7 +16,8 @@
     "frameworks": {
         "dnx452": {
             "dependencies": {
-                "EntityFramework": "6.1.3",
+                "EntityFramework.Core": "7.0.0-rc1-final",
+                "EntityFramework.Sqlite": "7.0.0-rc1-final",
 
                 "System.Linq.Parallel": "4.0.1-beta-*",
                 "System.Threading.Tasks": "4.0.11-beta-*"
